removed unused code from lamdaMailer
