+++ .gitignore
@@ -1,1 +1,2 @@
 secrets.json
+/depricated
