--- contact-form-api/getParamsFromUrl.js
@@ -1,8 +1,0 @@
-module.exports = (url) =>
-    decodeURIComponent(url)
-        .split('&')
-        .reduce((acc, curr) => {
-            const [key, value] = curr.split('=')
-            acc[key] = value
-            return acc
-        }, {})
