+++ contact-form-api/handler.js
@@ -1,38 +1,73 @@
-// 'use strict';
-/* module.exports.hello = async (event) => {
-  return {
-    statusCode: 200,
-    body: JSON.stringify(
-      {
-        message: 'Go Serverless v1.0! Your function executed successfully!',
-        input: event,
-      },
-      null,
-      2
-    ),
-  };
-}; */
+const aws = require('aws-sdk')
+const ses = new aws.SES()
 
-// define the options for email and domain
-const options = {
-  mySender: process.env.SENDER, // email address enabled in AWS SES in the AWS Console
-  myRecipient: process.env.RECIPIENT, // email address enabled in AWS SES in the AWS Console
-  myDomain: process.env.DOMAIN // the domain of my website or '*' to accept requests from any domain
+const mySender = process.env.SENDER; // email address 1 enabled in AWS SES in the AWS Console
+const myRecipient = process.env.RECIPIENT; // email address enabled in AWS SES in the AWS Console
+const myDomain = process.env.DOMAIN; // the domain of my website or '*' to accept requests from any domain
+
+function generateResponse(code, payload) {
+  return {
+    statusCode: code,
+    headers: {
+      'Access-Control-Allow-Origin': myDomain,
+      'Access-Control-Allow-Headers': 'x-requested-with',
+      'Access-Control-Allow-Credentials': true
+    },
+    body: JSON.stringify(payload)
+  }
+}
+function generateError(code, err) {
+  console.log(err)
+  return {
+    statusCode: code,
+    headers: {
+      'Access-Control-Allow-Origin': myDomain,
+      'Access-Control-Allow-Headers': 'x-requested-with',
+      'Access-Control-Allow-Credentials': true
+    },
+    body: JSON.stringify(err.message)
+  }
 }
 
-// initialize the function
-const { sendJSON, sendFormEncoded } = require('./lambdaMailer')(options);
+function generateEmailParamsFromJSON(body) {
+  const { email, name, content } = JSON.parse(body)
+  console.log(email, name, content)
+  if (!(email && name && content)) {
+    throw new Error('Missing parameters! Make sure to add parameters \'email\', \'name\', \'content\'.')
+  }
+
+  return {
+    Source: mySender,
+    Destination: { ToAddresses: [myRecipient] },
+    ReplyToAddresses: [email],
+    Message: {
+      Body: {
+        Text: {
+          Charset: 'UTF-8',
+          Data: `Message sent from email ${email} by ${name} \nContent: ${content}`
+        }
+      },
+      Subject: {
+        Charset: 'UTF-8',
+        Data: `New message from ${myDomain} portfolio contact form!`
+      }
+    }
+  }
+}
+
+async function sendJSON(event) {
+  try {
+    const emailParams = generateEmailParamsFromJSON(event.body)
+    const data = await ses.sendEmail(emailParams).promise()
+    return generateResponse(200, data)
+  } catch (err) {
+    return generateError(500, err)
+  }
+}
 
 // Content-Type: application/json
 // The event.body needs to be a JSON object with 3 properties
 // - email
 // - name
 // - content
-module.exports.sendJSON = sendJSON;
-
-// Content-Type: application/x-www-form-urlencoded
-// The event.body needs to a URI encoded string with 3 parameters
-// - email
-// - name
-// - content
-module.exports.sendFormEncoded = sendFormEncoded;
+module.exports.sendMail = sendJSON;
